# Conflicts:
#	src/main/java/space/obminyashka/items_exchange/controller/AuthController.java
#	src/main/java/space/obminyashka/items_exchange/service/JwtTokenService.java
#	src/main/java/space/obminyashka/items_exchange/service/impl/AuthServiceImpl.java

+++ src/main/java/space/obminyashka/items_exchange/service/CategoryService.java
@@ -85,5 +85,5 @@ public interface CategoryService {
      * @see Size.Clothing
      * @see Size.Shoes
      */
-    List<String> findSizesForCategory(int id);
+    List<String> findSizesForCategory(long id);
 }
